streamlining tests, removing scipy from requirements.
PiperOrigin-RevId: 424358509
