+++ ott/version.py
@@ -15,4 +15,4 @@
 
 """Current ott version."""
 
-__version__ = "0.1.20"
+__version__ = "0.2.0"
