+++ requirements.txt
@@ -3,6 +3,5 @@ jax>=0.1.67
 jaxlib>=0.1.47
 numpy>=1.18.4
 matplotlib>=2.0.1
-scipy>=1.0.0
 flax>=0.3.6
 optax>=0.0.9
