+++ setup.cfg
@@ -15,7 +15,7 @@ classifiers =
 
 [options]
 packages = find:
-python_requires = >=3.8
+python_requires = >=3.7
 
 [options.packages.find]
 exclude =
