+++ tests/core/sinkhorn_diff_precond_test.py
